Convert x86_Connect from extras to native-type example
