+++ platformio.ini
@@ -11,7 +11,13 @@ src_dir=.
 [common]
 lib_deps=
 build_flags=-Isrc
-src_filter=+<*> -<extras>
+build_src_filter=+<src/**/*.c> +<src/**/*.cpp>
+example_build_src_filter=+<examplesMain.cpp> ${common.build_src_filter}
+
+[env:ConSubNativeUnix]
+platform = native
+build_flags=${common.build_flags}
+build_src_filter=+<examples/ConSubNativeUnix/ConSubNativeUnix.cpp> ${common.build_src_filter}
 
 [env:PubSub_pro8MHzatmega328]
 platform=atmelavr
@@ -19,7 +25,7 @@ board=pro8MHzatmega328
 framework=arduino
 lib_deps=${common.lib_deps}
 		SoftwareSerial
-src_filter=${common.src_filter}
+build_src_filter=${common.example_build_src_filter}
 build_flags=${common.build_flags} -DMAIN_CPP_FILE=examples/PubSub/PubSub.ino
 
 [env:ConnectEthernetClient_pro8MHzatmega328]
@@ -29,7 +35,7 @@ framework=arduino
 lib_deps=${common.lib_deps}
 		SPI
 		Ethernet
-src_filter=${common.src_filter}
+build_src_filter=${common.example_build_src_filter}
 build_flags=${common.build_flags} -DMAIN_CPP_FILE=examples/ConnectEthernetClient/ConnectEthernetClient.ino
 
 [env:ConnectEsp8266WiFiClient_esp12e]
@@ -38,7 +44,7 @@ board=esp12e
 framework=arduino
 lib_deps=${common.lib_deps}
 		ESP8266WiFi
-src_filter=${common.src_filter}
+build_src_filter=${common.example_build_src_filter}
 build_flags=${common.build_flags} -DMAIN_CPP_FILE=examples/ConnectEsp8266WiFiClient/ConnectEsp8266WiFiClient.ino
 
 [env:LwtEsp8266WiFiClient_esp12e]
@@ -47,6 +53,5 @@ board=esp12e
 framework=arduino
 lib_deps=${common.lib_deps}
 		ESP8266WiFi
-src_filter=${common.src_filter}
+build_src_filter=${common.example_build_src_filter}
 build_flags=${common.build_flags} -DMAIN_CPP_FILE=examples/LwtEsp8266WiFiClient/LwtEsp8266WiFiClient.ino
-
